pyproject.toml
[project]
name = "myresume"
dynamic = ["version"]
description = "Create resumes from YAML description files"
readme = "README.md"
authors = [{name = "Albert Hopkins", email = "marduk@letterboxes.org"}]
license = {text = "GPLv3"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3"
]
requires-python = ">=3.11"
dependencies = [
    "Jinja2>=2.11.2",
    "PyYAML>=5.3.1",
    "weasyprint>=62.1",
    "html2text>=2025.4.15",
    "python-slugify>=8.0.4",
]

[project.urls]
homepage = "https://github.com/enku/myresume"
repository = "https://github.com/enku/myresume"

[project.scripts]
myresume = "myresume.cli:main"

[project.entry-points."myresume.themes"]
default = "myresume.themes.default"

[build-system]
requires = [
    "pdm-pep517>=1.0.0"
]

build-backend = "pdm.pep517.api"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "coverage>=6.4.4",
    "black>=22.8.0",
    "mypy>=0.971",
    "pylint>=2.15.3",
    "types-PyYAML>=6.0.12.12",
]

[tool.black]
skip-magic-trailing-comma = true

[tool.coverage.run]
source = ["myresume", "tests"]
branch = true
omit = ["tests/__main__.py"]

[tool.pylint.master]
load-plugins = "pylint.extensions.bad_builtin"
bad-functions = "print"

[tool.pylint.variables]
init-import = "yes"
